feat : rename Page
