+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { AuthPage } from './authPage/Auth'
+export { AuthPage } from './authPage/AuthPage'
 export { Main } from './main/Main'
 export { ProfilePage } from './ProfilePage/ProfilePage'
 export { Course } from './course/Course'
